requirements.txt
playsound==1.2.2
pillow==11.3.0